+++ BE/src/main/java/ifive/idrop/controller/CommonController.java
@@ -4,12 +4,12 @@ import ifive.idrop.annotation.Login;
 import ifive.idrop.dto.response.BaseResponse;
 import ifive.idrop.dto.response.DriverDetailResponse;
 import ifive.idrop.driver.domain.Driver;
-import ifive.idrop.entity.Parent;
+import ifive.idrop.parent.domain.Parent;
 import ifive.idrop.entity.Users;
 import ifive.idrop.exception.CommonException;
 import ifive.idrop.exception.ErrorCode;
 import ifive.idrop.driver.service.DriverService;
-import ifive.idrop.service.ParentService;
+import ifive.idrop.parent.service.ParentService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
