+++ BE/src/main/java/ifive/idrop/driver/service/DriverService.java
@@ -12,6 +12,7 @@ import ifive.idrop.exception.ErrorCode;
 import ifive.idrop.fcm.AlarmMessage;
 import ifive.idrop.fcm.NotificationUtill;
 import ifive.idrop.driver.repository.DriverRepository;
+import ifive.idrop.parent.domain.Parent;
 import ifive.idrop.repository.NotificationRepository;
 import ifive.idrop.repository.PickUpRepository;
 import ifive.idrop.util.RequestSchedule;
