+++ BE/src/main/java/ifive/idrop/dto/response/DriverSubscribeInfoResponse.java
@@ -1,6 +1,7 @@
 package ifive.idrop.dto.response;
 
 import ifive.idrop.entity.*;
+import ifive.idrop.parent.domain.Parent;
 import lombok.Builder;
 import lombok.Getter;
 import org.json.simple.JSONObject;
