+++ BE/src/main/java/ifive/idrop/entity/PickUp.java
@@ -1,6 +1,7 @@
 package ifive.idrop.entity;
 
 import ifive.idrop.driver.domain.Driver;
+import ifive.idrop.parent.domain.Parent;
 import jakarta.persistence.*;
 import lombok.*;
 import lombok.extern.slf4j.Slf4j;
