+++ BE/src/main/java/ifive/idrop/entity/PickUpInfo.java
@@ -1,6 +1,7 @@
 package ifive.idrop.entity;
 
 import ifive.idrop.driver.domain.Driver;
+import ifive.idrop.parent.domain.Parent;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.AllArgsConstructor;
