+++ BE/src/main/java/ifive/idrop/parent/domain/Parent.java
@@ -1,5 +1,7 @@
-package ifive.idrop.entity;
+package ifive.idrop.parent.domain;
 
+import ifive.idrop.entity.Child;
+import ifive.idrop.entity.Users;
 import jakarta.persistence.*;
 import lombok.Getter;
