+++ BE/src/main/java/ifive/idrop/websocket/location/LocationWebSocketHandler.java
@@ -8,6 +8,7 @@ import ifive.idrop.exception.ErrorCode;
 import ifive.idrop.filter.AuthenticateUser;
 import ifive.idrop.filter.VerifyUserFilter;
 import ifive.idrop.jwt.JwtProvider;
+import ifive.idrop.parent.domain.Parent;
 import ifive.idrop.repository.UserRepository;
 import ifive.idrop.util.CustomObjectMapper;
 import ifive.idrop.websocket.PickUpInfoRepository;
